requirements.txt
lorem>=0.1.1
munch>=4.0.0
pykube-ng>=23.6.0
pdoc>=15.0.0
